--- node/src/cli/cli_standalone.rs
@@ -1,76 +1,0 @@
-use sc_cli::{ChainSpec, RunCmd, RuntimeVersion, SubstrateCli};
-use structopt::StructOpt;
-
-#[derive(Debug, StructOpt)]
-pub struct Cli {
-    #[structopt(subcommand)]
-    pub subcommand: Option<Subcommand>,
-
-    #[structopt(flatten)]
-    pub run: RunCmd,
-}
-
-#[derive(Debug, StructOpt)]
-pub enum Subcommand {
-    /// The custom benchmark subcommmand benchmarking runtime pallets.
-    #[structopt(name = "benchmark", about = "Benchmark runtime pallets.")]
-    Benchmark(frame_benchmarking_cli::BenchmarkCmd),
-
-    /// Build a chain specification.
-    BuildSpec(sc_cli::BuildSpecCmd),
-
-    /// Validate blocks.
-    CheckBlock(sc_cli::CheckBlockCmd),
-
-    /// Export blocks.
-    ExportBlocks(sc_cli::ExportBlocksCmd),
-
-    /// Export the state of a given block into a chain spec.
-    ExportState(sc_cli::ExportStateCmd),
-
-    /// Import blocks.
-    ImportBlocks(sc_cli::ImportBlocksCmd),
-
-    /// Key management cli utilities
-    Key(sc_cli::KeySubcommand),
-
-    /// Remove the whole chain.
-    PurgeChain(sc_cli::PurgeChainCmd),
-
-    /// Revert the chain to a previous state.
-    Revert(sc_cli::RevertCmd),
-}
-
-impl SubstrateCli for Cli {
-    fn author() -> String {
-        env!("CARGO_PKG_AUTHORS").into()
-    }
-
-    fn copyright_start_year() -> i32 {
-        2021
-    }
-
-    fn description() -> String {
-        env!("CARGO_PKG_DESCRIPTION").into()
-    }
-
-    fn impl_name() -> String {
-        "Zeitgeist Node".into()
-    }
-
-    fn impl_version() -> String {
-        env!("SUBSTRATE_CLI_IMPL_VERSION").into()
-    }
-
-    fn load_spec(&self, id: &str) -> Result<Box<dyn sc_service::ChainSpec>, String> {
-        crate::cli::load_spec(id)
-    }
-
-    fn native_runtime_version(_: &Box<dyn ChainSpec>) -> &'static RuntimeVersion {
-        &zeitgeist_runtime::VERSION
-    }
-
-    fn support_url() -> String {
-        "support.anonymous.an".into()
-    }
-}
